v0.05 -  Creating Roles
This time we continue to create some more functions so that we can move forward in building our application. We have created two types of functions the first one is related to encryption and decryption of text with a defined key, the other is a function to encrypt and encrypt the text using an anonymous key defined by the user at the time of compression.